fix(tax): missing soft-delete cols (periods) + int4 total_count decode bug (#162) (#164)
Two independent bugs in the tax module, both in the same FromRow decode
root-cause class tracked by #162:

Bug 1 — tax_periods (migration 012) was never given deleted_at/deleted_by,
but TaxPeriod uses impl_soft_deletable! and the postgres repo SELECTs them
and filters WHERE deleted_at IS NULL on every read. migration 020 missed
tax_periods. Result: GET /api/v1/tax/periods (list, find_by_id, mutations)
500'd with "column deleted_at does not exist".

  Fix: migration 040 adds deleted_at TIMESTAMPTZ, deleted_by BIGINT
  REFERENCES users(id) ON DELETE SET NULL, and the partial index
  idx_tax_periods_deleted_at ... WHERE deleted_at IS NULL. Idempotent +
  down migration. Registered in BOTH the MIGRATIONS (up) and reverse-ordered
  DOWN_MIGRATIONS arrays in src/db/pool.rs — the static migration runner
  has no directory globbing, so an unregistered .sql file is never applied
  (caught by adversarial review).

Bug 2 — every single-row query (find_by_id, create/update/restore
RETURNING — 11 sites) used `0 as total_count`. Postgres types the literal
`0` as integer (int4), but TaxRateRow/TaxPeriodRow declare
total_count: Option<i64> (int8) → sqlx type-oid mismatch → 500. The list
path used COUNT(*) OVER() (int8) so GET /tax/rates and GET /tax/periods
lists worked and HID the bug — the "list suite hides decode bugs" pattern.

  Fix: `0 as total_count` -> `0::bigint as total_count` (all 11 sites).

Live verification (rebuilt binary on :8081):
  GET /tax/periods        500 -> 200   GET /tax/rates/{id}   500 -> 200
  GET /tax/periods/{id}  500 -> 200   GET /tax/rates/999999 404 -> 404
Full hurl suite (56 files) on :8081 green. cargo test tax (56 lib tests),
clippy -D warnings, fmt --check all clean.

Hurl (tests/hurl/28_tax.hurl): periods list fence flipped 500->200+envelope;
added a [Captures] of items[0].id + GET /tax/rates/{id} -> 200 assertion that
actually fences the int4 find_by_id decode fix (the list path never did).
README known-broken table: removed the 28_tax row.

Codebase-wide follow-up (out of scope, noted for a sweep): the identical
`0 as total_count` int4 bug exists in cost_center, edefter, archive, efatura
repositories — their get-by-id/mutation paths are similarly broken.

+++ turerp/src/domain/tax/postgres_repository.rs
@@ -204,7 +204,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
             RETURNING "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(tenant_id)
@@ -227,7 +227,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             r#"
             SELECT "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM tax_rates
             WHERE id = $1 AND tenant_id = $2 AND deleted_at IS NULL
             "#
@@ -326,7 +326,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             r#"
             SELECT "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM tax_rates
             WHERE tenant_id = $1
               AND tax_type = $2
@@ -365,7 +365,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             WHERE id = $6 AND tenant_id = $7 AND deleted_at IS NULL
             RETURNING "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(update.rate)
@@ -433,7 +433,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             WHERE id = $1 AND tenant_id = $2 AND deleted_at IS NOT NULL
             RETURNING "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(id)
@@ -450,7 +450,7 @@ impl TaxRateRepository for PostgresTaxRateRepository {
             r#"
             SELECT "#,
             TAX_RATE_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM tax_rates
             WHERE tenant_id = $1 AND deleted_at IS NOT NULL
             ORDER BY deleted_at DESC
@@ -532,7 +532,7 @@ impl TaxPeriodRepository for PostgresTaxPeriodRepository {
             VALUES ($1, $2, $3, $4)
             RETURNING "#,
             TAX_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(tenant_id)
@@ -551,7 +551,7 @@ impl TaxPeriodRepository for PostgresTaxPeriodRepository {
             r#"
             SELECT "#,
             TAX_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM tax_periods
             WHERE id = $1 AND tenant_id = $2 AND deleted_at IS NULL
             "#
@@ -662,7 +662,7 @@ impl TaxPeriodRepository for PostgresTaxPeriodRepository {
             WHERE id = $10 AND tenant_id = $11 AND deleted_at IS NULL
             RETURNING "#,
             TAX_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(&tax_type_str)
@@ -756,7 +756,7 @@ impl TaxPeriodRepository for PostgresTaxPeriodRepository {
             WHERE id = $1 AND tenant_id = $2 AND deleted_at IS NOT NULL
             RETURNING "#,
             TAX_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             "#
         ))
         .bind(id)
@@ -773,7 +773,7 @@ impl TaxPeriodRepository for PostgresTaxPeriodRepository {
             r#"
             SELECT "#,
             TAX_PERIOD_COLUMNS!(),
-            r#", 0 as total_count
+            r#", 0::bigint as total_count
             FROM tax_periods
             WHERE tenant_id = $1 AND deleted_at IS NOT NULL
             ORDER BY deleted_at DESC
